requirements.txt
blpapi>=3.19.1
pandas>=1.5.0
python-dotenv>=1.0.0
O365>=2.0.26
openpyxl>=3.1.2
python-docx>=0.8.11 